+++ docs/changelog.rst
@@ -3,7 +3,8 @@ Changelog
 
 Version 1.2.11
 ------------------
-增加 `MODEL_SET_VERBOSE_NAME_TO_DB_COMMENT`配置， 当 model 的 field 未设置 `db_comment` 时, 自动将 `db_comment` 设置为 `verbose_name` 的值
++ 增加 `MODEL_SET_VERBOSE_NAME_TO_DB_COMMENT`配置， 当 model 的 field 未设置 `db_comment` 时, 自动将 `db_comment` 设置为 `verbose_name` 的值
++ Fixed python3.11中 `Constants` 排重逻辑不正确
 
 Version 1.2.10
 ------------------